capability types organized
