+++ src/types/container/index.ts
@@ -19,11 +19,29 @@ export type PieceCloseAction = {
     };
 };
 
+export const SERVER_INNER_PIECE_CLICK = 'SERVER_INNER_PIECE_CLICK';
+export type ExitContainerRequestAction = {
+    type: typeof SERVER_INNER_PIECE_CLICK;
+    payload: {
+        selectedPiece: PieceType;
+        containerPiece: PieceType;
+    };
+};
+
 export const INNER_PIECE_CLICK_ACTION = 'INNER_PIECE_CLICK_ACTION';
 export type ExitContainerAction = {
     type: typeof INNER_PIECE_CLICK_ACTION;
     payload: {
         gameboardPieces: GameboardPiecesDataType;
+        selectedPiece: PieceType;
+        containerPiece: PieceType;
+    };
+};
+
+export const SERVER_OUTER_PIECE_CLICK = 'SERVER_OUTER_PIECE_CLICK';
+export type EnterContainerRequestAction = {
+    type: typeof SERVER_OUTER_PIECE_CLICK;
+    payload: {
         selectedPiece: PieceType;
         containerPiece: PieceType;
     };
@@ -39,24 +57,6 @@ export type EnterContainerAction = {
     };
 };
 
-export const INNER_TRANSPORT_PIECE_CLICK_ACTION = 'INNER_TRANSPORT_PIECE_CLICK_ACTION';
-export type ExitTransportContainerAction = {
-    type: typeof INNER_TRANSPORT_PIECE_CLICK_ACTION;
-    payload: {
-        selectedPiece: PieceType;
-        containerPiece: PieceType;
-    };
-};
-
-export const SERVER_INNER_PIECE_CLICK = 'SERVER_INNER_PIECE_CLICK';
-export type ExitContainerRequestAction = {
-    type: typeof SERVER_INNER_PIECE_CLICK;
-    payload: {
-        selectedPiece: PieceType;
-        containerPiece: PieceType;
-    };
-};
-
 export const SERVER_INNER_TRANSPORT_PIECE_CLICK = 'SERVER_INNER_TRANSPORT_PIECE_CLICK';
 export type ExitTransportContainerRequestAction = {
     type: typeof SERVER_INNER_TRANSPORT_PIECE_CLICK;
@@ -67,9 +67,9 @@ export type ExitTransportContainerRequestAction = {
     };
 };
 
-export const SERVER_OUTER_PIECE_CLICK = 'SERVER_OUTER_PIECE_CLICK';
-export type EnterContainerRequestAction = {
-    type: typeof SERVER_OUTER_PIECE_CLICK;
+export const INNER_TRANSPORT_PIECE_CLICK_ACTION = 'INNER_TRANSPORT_PIECE_CLICK_ACTION';
+export type ExitTransportContainerAction = {
+    type: typeof INNER_TRANSPORT_PIECE_CLICK_ACTION;
     payload: {
         selectedPiece: PieceType;
         containerPiece: PieceType;
