+++ src/types/planning/index.ts
@@ -14,20 +14,26 @@ export type PreventPlanAction = {
     payload: {};
 };
 
+export const PLANNING_SELECT = 'PLANNING_SELECT';
+export type PlanningSelectAction = {
+    type: typeof PLANNING_SELECT;
+    payload: {
+        selectedPositionId: GameboardMetaState['selectedPosition'];
+    };
+};
+
+export const UNDO_MOVE = 'UNDO_MOVE';
+export type UndoMoveAction = {
+    type: typeof UNDO_MOVE;
+    payload: {};
+};
+
 export const SERVER_CONFIRM_PLAN = 'SERVER_CONFIRM_PLAN';
 export type ConfirmPlanRequestAction = {
     type: typeof SERVER_CONFIRM_PLAN;
     payload: {
         pieceId: PieceType['pieceId'];
         plan: LIST_ALL_POSITIONS_TYPE[];
-    };
-};
-
-export const PLANNING_SELECT = 'PLANNING_SELECT';
-export type PlanningSelectAction = {
-    type: typeof PLANNING_SELECT;
-    payload: {
-        selectedPositionId: GameboardMetaState['selectedPosition'];
     };
 };
 
@@ -40,14 +46,6 @@ export type ConfirmPlanAction = {
     };
 };
 
-export const DELETE_PLAN = 'DELETE_PLAN';
-export type DeletePlanAction = {
-    type: typeof DELETE_PLAN;
-    payload: {
-        pieceId: PieceType['pieceId'];
-    };
-};
-
 export const SERVER_DELETE_PLAN = 'SERVER_DELETE_PLAN';
 export type DeletePlanRequestAction = {
     type: typeof SERVER_DELETE_PLAN;
@@ -56,8 +54,10 @@ export type DeletePlanRequestAction = {
     };
 };
 
-export const UNDO_MOVE = 'UNDO_MOVE';
-export type UndoMoveAction = {
-    type: typeof UNDO_MOVE;
-    payload: {};
+export const DELETE_PLAN = 'DELETE_PLAN';
+export type DeletePlanAction = {
+    type: typeof DELETE_PLAN;
+    payload: {
+        pieceId: PieceType['pieceId'];
+    };
 };
