+++ src/types/refuel/index.ts
@@ -10,27 +10,6 @@ export type RefuelOpenAction = {
     };
 };
 
-export const UNDO_FUEL_SELECTION = 'UNDO_FUEL_SELECTION';
-export type UndoFuelSelectionAction = {
-    type: typeof UNDO_FUEL_SELECTION;
-    payload: {
-        aircraftPiece: any;
-        aircraftPieceIndex: number;
-    };
-};
-
-export const REFUEL_RESULTS = 'REFUEL_RESULTS';
-export type FuelResultsAction = {
-    type: typeof REFUEL_RESULTS;
-    payload: {
-        fuelUpdates: {
-            pieceId: PieceType['pieceId'];
-            piecePositionId: PieceType['piecePositionId'];
-            newFuel: PieceType['pieceFuel'];
-        }[];
-    };
-};
-
 export const REFUELPOPUP_MINIMIZE_TOGGLE = 'REFUELPOPUP_MINIMIZE_TOGGLE';
 export type RefuelPopupToggleAction = {
     type: typeof REFUELPOPUP_MINIMIZE_TOGGLE;
@@ -55,11 +34,32 @@ export type AircraftClickAction = {
     };
 };
 
+export const UNDO_FUEL_SELECTION = 'UNDO_FUEL_SELECTION';
+export type UndoFuelSelectionAction = {
+    type: typeof UNDO_FUEL_SELECTION;
+    payload: {
+        aircraftPiece: any;
+        aircraftPieceIndex: number;
+    };
+};
+
 export const SERVER_CONFIRM_FUEL_SELECTION = 'SERVER_CONFIRM_FUEL_SELECTION';
 export type ConfirmFuelSelectionRequestAction = {
     type: typeof SERVER_CONFIRM_FUEL_SELECTION;
     payload: {
         aircraft: RefuelState['aircraft'];
         tankers: RefuelState['tankers'];
+    };
+};
+
+export const REFUEL_RESULTS = 'REFUEL_RESULTS';
+export type FuelResultsAction = {
+    type: typeof REFUEL_RESULTS;
+    payload: {
+        fuelUpdates: {
+            pieceId: PieceType['pieceId'];
+            piecePositionId: PieceType['piecePositionId'];
+            newFuel: PieceType['pieceFuel'];
+        }[];
     };
 };
